+++ projects/ngx-xchange-ui/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.0.15 - 08/12/2023
+* (Refactor) Refactor details, two way open attribute
+
+# 0.0.14 - 08/12/2023
+* (Add) Create rate component
+
 # 0.0.13 - 06/12/2023
 * (Remove) Symbol ' on button view component